fix to last
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/branch/alpine/release-2.5/alpine@1974 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ tran/spam/main/com/metavize/tran/spam/SpamProtoConfig.java
@@ -118,7 +118,7 @@ public abstract class SpamProtoConfig implements Serializable
 
     public void setStrengthByName(String strengthEnum)
     {
-        Integer str = scanStrengthEnumeration.get(strengthEnum);
+        Integer str = (Integer) scanStrengthEnumeration.get(strengthEnum);
         if (str != null)
             strength = str.intValue();
     }
